+++ src/LibHac/Tools/Npdm/ApplicationType.cs
@@ -1,4 +1,4 @@
-namespace LibHac.Npdm;
+namespace LibHac.Tools.Npdm;
 
 internal enum ApplicationType
 {
