+++ src/LibHac/Tools/Npdm/FsPermissionRw.cs
@@ -1,5 +1,5 @@
 // ReSharper disable InconsistentNaming
-namespace LibHac.Npdm;
+namespace LibHac.Tools.Npdm;
 
 public enum FsPermissionRw : ulong
 {
