+++ src/LibHac/Tools/Npdm/ServiceAccessControl.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.IO;
 using LibHac.Common;
 
-namespace LibHac.Npdm;
+namespace LibHac.Tools.Npdm;
 
 public class ServiceAccessControl
 {
