+++ src/LibHac/Tools/Npdm/SvcName.cs
@@ -1,4 +1,4 @@
-namespace LibHac.Npdm;
+namespace LibHac.Tools.Npdm;
 
 public enum SvcName
 {
