+++ src/hactoolnet/ProcessNca.cs
@@ -6,10 +6,10 @@ using LibHac.Fs;
 using LibHac.Fs.Fsa;
 using LibHac.Fs.Impl;
 using LibHac.FsSystem;
-using LibHac.Npdm;
 using LibHac.Tools.Fs;
 using LibHac.Tools.FsSystem;
 using LibHac.Tools.FsSystem.NcaUtils;
+using LibHac.Tools.Npdm;
 using static hactoolnet.Print;
 
 namespace hactoolnet;
