+++ tests/LibHac.Tests/Fs/LayeredFileSystemTests.cs
@@ -2,8 +2,8 @@
 using LibHac.Common;
 using LibHac.Fs;
 using LibHac.Fs.Fsa;
-using LibHac.FsSystem;
 using LibHac.Tools.Fs;
+using LibHac.Tools.FsSystem;
 using LibHac.Util;
 using Xunit;
